requirements.txt
torch>=2.0.0
torchvision
transformers==4.51.3
Pillow
param
scipy
black
einops
matplotlib